imagetools has been split out into separate files, grouping related functions.
